# packages/api-contract/openapi.json
{
  "components": {
    "schemas": {
...
      "WorkspaceCollectionResponse": {
        "additionalProperties": false,
        "properties": {
          "workspaces": {
            "items": {
              "$ref": "#/components/schemas/WorkspaceResponse"
            },
            "title": "Workspaces",
            "type": "array"
          }
        },
        "required": ["workspaces"],
        "title": "WorkspaceCollectionResponse",
        "type": "object"
      },
...
      "WorkspaceResponse": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/WorkspaceRoleName"
          }
        },
        "required": ["id", "name", "role"],
        "title": "WorkspaceResponse",
        "type": "object"
      },
      "WorkspaceRoleName": {
        "enum": ["owner", "admin", "member", "viewer"],
        "type": "string"
      }
    },
    "securitySchemes": {
      "AccessToken": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Industry Intelligence Platform API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/workspaces": {
      "get": {
        "operationId": "list_workspaces_api_v1_workspaces_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceCollectionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "additionalProperties": false,
                  "description": "Stable, non-sensitive error contract returned by the API.",
                  "properties": {
                    "code": {
                      "title": "Code",
                      "type": "string"
                    },
                    "detail": {
                      "title": "Detail",
                      "type": "string"
                    },
                    "status": {
                      "title": "Status",
                      "type": "integer"
                    },
                    "title": {
                      "title": "Title",
                      "type": "string"
                    },
                    "trace_id": {
                      "title": "Trace Id",
                      "type": "string"
                    },
                    "type": {
                      "title": "Type",
                      "type": "string"
                    }
                  },
                  "required": ["type", "title", "status", "detail", "code", "trace_id"],
                  "title": "ProblemDetails",
                  "type": "object"
                }
              }
            },
            "description": "Invalid authenticated session"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "additionalProperties": false,
                  "description": "Stable, non-sensitive error contract returned by the API.",
                  "properties": {
                    "code": {
                      "title": "Code",
                      "type": "string"
                    },
                    "detail": {
                      "title": "Detail",
                      "type": "string"
                    },
                    "status": {
                      "title": "Status",
                      "type": "integer"
                    },
                    "title": {
                      "title": "Title",
                      "type": "string"
                    },
                    "trace_id": {
                      "title": "Trace Id",
                      "type": "string"
                    },
                    "type": {
                      "title": "Type",
                      "type": "string"
                    }
                  },
                  "required": ["type", "title", "status", "detail", "code", "trace_id"],
                  "title": "ProblemDetails",
                  "type": "object"
                }
              }
            },
            "description": "Internal server error"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "additionalProperties": false,
                  "description": "Stable, non-sensitive error contract returned by the API.",
                  "properties": {
                    "code": {
                      "title": "Code",
                      "type": "string"
                    },
                    "detail": {
                      "title": "Detail",
                      "type": "string"
                    },
                    "status": {
                      "title": "Status",
                      "type": "integer"
                    },
                    "title": {
                      "title": "Title",
                      "type": "string"
                    },
                    "trace_id": {
                      "title": "Trace Id",
                      "type": "string"
                    },
                    "type": {
                      "title": "Type",
                      "type": "string"
                    }
                  },
                  "required": ["type", "title", "status", "detail", "code", "trace_id"],
                  "title": "ProblemDetails",
                  "type": "object"
                }
              }
            },
            "description": "Workspace service temporarily unavailable"
          }
        },
        "security": [
          {
            "AccessToken": []
          }
        ],
        "summary": "List Workspaces",
        "tags": ["workspaces"]
      }
    },
...
  }
}